Moved entity.h and adaptor_common.h into the main include path.
