+++ include/qpid/dispatch/entity.h
