+++ tests/cpp/cpp_system/test_listener_startup.cpp
@@ -29,7 +29,7 @@
 
 extern "C" {
 #include "adaptors/amqp/qd_listener.h"
-#include "entity.h"
+#include "qpid/dispatch/entity.h"
     qd_listener_t *qd_dispatch_configure_listener(qd_dispatch_t *qd, qd_entity_t *entity);
     void qd_connection_manager_delete_listener(qd_dispatch_t *qd, void *impl);
 }
